network: cancel timeouts when releasing the timer
Otherwise these timers will never be canceled and may be executed after the
object ist deleted.
Also drop the check in OnTimeout(). The timeout is canceled, so OnTimeout()
is never called for a released timer.

This is a followup patch for 85d82e90857d509c7da2ec7fa95feae233848479 which
did not fix all possible race conditions.