Create a Client::Parameters struct to avoid passing IEs
IWD doesn't accept or export raw IEs on the DBus interface, instead it
takes individual parameter values and only a subset of all the ones that
the IEs contain.  Make the connman-specific Client class build the IE
locally from an abstract parameters struct.  For now the struct contains
the values that source-app.cpp and sink-app.cpp use, if they need to
indicate support for any of the extended WFD capabilities or other
parameters, this struct will have to be extended too.

This might also be a step in supporting Miracast over Ethernet (MICE?)
if ever needed -- I think this is the name used for the R2-specific
connection method where parameters are exposed as mDNS records.