p2p: Add MultiClient
To avoid forcing the source/sink apps to select between one client or
another, add the MultiClient class, implementing P2P::Client, so that
it can be a drop-in replacement of ConnmanClient that does the right
thing depending on which of the two services are available on DBus.
Under the hood it hold objects of both ConnmandClient and IWDClient and
whichever detects its corresponding daemon is available for use,
forwards any callbacks through MultiClient up to the app.